refactor(blockchain): remove Provider adoptor
