--- src/connectors/blockchain/contract/index.ts
@@ -1,1 +1,0 @@
-export * from './curation'
